chore: Rename test DB initializers to include 'Outbox' in names
Renamed all test database initializer classes to explicitly include 'Outbox' in their names (e.g., EntityFrameworkOutboxInitializer, SqlServerAdoNetOutboxInitializer, etc.). Updated all test attributes and usages to reference the new class names for improved clarity and intent in Outbox-related integration tests.
