+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/PostgreSqlAdoNetOutboxInitializer.cs
@@ -2,32 +2,32 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
-using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using NetEvolve.Pulse;
 using NetEvolve.Pulse.Extensibility;
 using NetEvolve.Pulse.Extensibility.Outbox;
 using NetEvolve.Pulse.Outbox;
+using Npgsql;
 
 [SuppressMessage(
     "Security",
     "CA2100:Review SQL queries for security vulnerabilities",
     Justification = "SQL is read from a script file with schema and table names substituted from validated OutboxOptions properties."
 )]
-public sealed partial class SqlServerAdoNetInitializer : IDatabaseInitializer
+public sealed partial class PostgreSqlAdoNetOutboxInitializer : IDatabaseInitializer
 {
     private static readonly string _scriptPath = Path.Combine(
         AppContext.BaseDirectory,
         "Scripts",
-        "SqlServer",
+        "PostgreSql",
         "OutboxMessage.sql"
     );
 
     public void Configure(IMediatorBuilder mediatorBuilder, IDatabaseServiceFixture databaseService)
     {
         ArgumentNullException.ThrowIfNull(databaseService);
-        _ = mediatorBuilder.AddSqlServerOutbox(databaseService.ConnectionString);
+        _ = mediatorBuilder.AddPostgreSqlOutbox(databaseService.ConnectionString);
     }
 
     public async ValueTask CreateDatabaseAsync(IServiceProvider serviceProvider, CancellationToken cancellationToken)
@@ -46,38 +46,34 @@ public sealed partial class SqlServerAdoNetInitializer : IDatabaseInitializer
 
         var script = await File.ReadAllTextAsync(_scriptPath, cancellationToken).ConfigureAwait(false);
 
-        // Remove SQLCMD-specific variable declarations (not valid T-SQL)
+        // Remove psql-specific variable declarations (not valid SQL)
         script = SearchSetVar().Replace(script, string.Empty);
 
-        // Substitute SQLCMD variables with actual values
+        // Substitute psql variables with actual values.
+        // PostgreSQL script uses :schema_name and :table_name as placeholders.
+        // The placeholders appear both unquoted (e.g., CREATE SCHEMA :schema_name)
+        // and within quotes (e.g., ":schema_name".":table_name").
+        // We replace all occurrences with the actual values directly.
         script = script
-            .Replace("$(SchemaName)", schema, StringComparison.Ordinal)
-            .Replace("$(TableName)", tableName, StringComparison.Ordinal);
+            .Replace(":schema_name", schema, StringComparison.Ordinal)
+            .Replace(":table_name", tableName, StringComparison.Ordinal);
 
-        // Split on GO (on its own line) and execute each batch independently
-        var batches = SearchGoStatements().Split(script);
-
-        await using var connection = new SqlConnection(connectionString);
-        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
-
-        foreach (var batch in batches)
+        var connection = new NpgsqlConnection(connectionString);
+        await using (connection.ConfigureAwait(false))
         {
-            var trimmed = batch.Trim();
-            if (string.IsNullOrWhiteSpace(trimmed))
-            {
-                continue;
-            }
+            await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
 
-            await using var command = new SqlCommand(trimmed, connection);
+            await using var command = new NpgsqlCommand(script, connection);
             _ = await command.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
         }
     }
 
-    public void Initialize(IServiceCollection services, IDatabaseServiceFixture databaseService) { }
+    public void Initialize(IServiceCollection services, IDatabaseServiceFixture databaseService)
+    {
+        // No additional service initialization required for ADO.NET outbox tests.
+        // The Configure method handles all necessary service registrations.
+    }
 
-    [GeneratedRegex(@"^:setvar\s+\w+\s+.*$", RegexOptions.Multiline, 10000)]
+    [GeneratedRegex(@"^\\set\s+\w+\s+.*$", RegexOptions.Multiline, 10000)]
     private static partial Regex SearchSetVar();
-
-    [GeneratedRegex(@"^\s*GO\s*$", RegexOptions.IgnoreCase | RegexOptions.Multiline, 10000)]
-    private static partial Regex SearchGoStatements();
 }
