+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/SqlServerAdoNetOutboxInitializer.cs
@@ -2,32 +2,32 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
+using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using NetEvolve.Pulse;
 using NetEvolve.Pulse.Extensibility;
 using NetEvolve.Pulse.Extensibility.Outbox;
 using NetEvolve.Pulse.Outbox;
-using Npgsql;
 
 [SuppressMessage(
     "Security",
     "CA2100:Review SQL queries for security vulnerabilities",
     Justification = "SQL is read from a script file with schema and table names substituted from validated OutboxOptions properties."
 )]
-public sealed partial class PostgreSqlAdoNetInitializer : IDatabaseInitializer
+public sealed partial class SqlServerAdoNetOutboxInitializer : IDatabaseInitializer
 {
     private static readonly string _scriptPath = Path.Combine(
         AppContext.BaseDirectory,
         "Scripts",
-        "PostgreSql",
+        "SqlServer",
         "OutboxMessage.sql"
     );
 
     public void Configure(IMediatorBuilder mediatorBuilder, IDatabaseServiceFixture databaseService)
     {
         ArgumentNullException.ThrowIfNull(databaseService);
-        _ = mediatorBuilder.AddPostgreSqlOutbox(databaseService.ConnectionString);
+        _ = mediatorBuilder.AddSqlServerOutbox(databaseService.ConnectionString);
     }
 
     public async ValueTask CreateDatabaseAsync(IServiceProvider serviceProvider, CancellationToken cancellationToken)
@@ -46,34 +46,38 @@ public sealed partial class PostgreSqlAdoNetInitializer : IDatabaseInitializer
 
         var script = await File.ReadAllTextAsync(_scriptPath, cancellationToken).ConfigureAwait(false);
 
-        // Remove psql-specific variable declarations (not valid SQL)
+        // Remove SQLCMD-specific variable declarations (not valid T-SQL)
         script = SearchSetVar().Replace(script, string.Empty);
 
-        // Substitute psql variables with actual values.
-        // PostgreSQL script uses :schema_name and :table_name as placeholders.
-        // The placeholders appear both unquoted (e.g., CREATE SCHEMA :schema_name)
-        // and within quotes (e.g., ":schema_name".":table_name").
-        // We replace all occurrences with the actual values directly.
+        // Substitute SQLCMD variables with actual values
         script = script
-            .Replace(":schema_name", schema, StringComparison.Ordinal)
-            .Replace(":table_name", tableName, StringComparison.Ordinal);
+            .Replace("$(SchemaName)", schema, StringComparison.Ordinal)
+            .Replace("$(TableName)", tableName, StringComparison.Ordinal);
 
-        var connection = new NpgsqlConnection(connectionString);
-        await using (connection.ConfigureAwait(false))
+        // Split on GO (on its own line) and execute each batch independently
+        var batches = SearchGoStatements().Split(script);
+
+        await using var connection = new SqlConnection(connectionString);
+        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+
+        foreach (var batch in batches)
         {
-            await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+            var trimmed = batch.Trim();
+            if (string.IsNullOrWhiteSpace(trimmed))
+            {
+                continue;
+            }
 
-            await using var command = new NpgsqlCommand(script, connection);
+            await using var command = new SqlCommand(trimmed, connection);
             _ = await command.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
         }
     }
 
-    public void Initialize(IServiceCollection services, IDatabaseServiceFixture databaseService)
-    {
-        // No additional service initialization required for ADO.NET outbox tests.
-        // The Configure method handles all necessary service registrations.
-    }
+    public void Initialize(IServiceCollection services, IDatabaseServiceFixture databaseService) { }
 
-    [GeneratedRegex(@"^\\set\s+\w+\s+.*$", RegexOptions.Multiline, 10000)]
+    [GeneratedRegex(@"^:setvar\s+\w+\s+.*$", RegexOptions.Multiline, 10000)]
     private static partial Regex SearchSetVar();
+
+    [GeneratedRegex(@"^\s*GO\s*$", RegexOptions.IgnoreCase | RegexOptions.Multiline, 10000)]
+    private static partial Regex SearchGoStatements();
 }
