+++ tests/NetEvolve.Pulse.Tests.Integration/Outbox/SQLiteAdoNetOutboxTests.cs
@@ -3,7 +3,9 @@ namespace NetEvolve.Pulse.Tests.Integration.Outbox;
 using NetEvolve.Extensions.TUnit;
 using NetEvolve.Pulse.Tests.Integration.Internals;
 
-[ClassDataSource<SQLiteDatabaseServiceFixture, SQLiteAdoNetInitializer>(Shared = [SharedType.None, SharedType.None])]
+[ClassDataSource<SQLiteDatabaseServiceFixture, SQLiteAdoNetOutboxInitializer>(
+    Shared = [SharedType.None, SharedType.None]
+)]
 [TestGroup("SQLite")]
 [TestGroup("AdoNet")]
 [InheritsTests]
